Allow remote access from SSM Session manager.
This will not completely fix #620, because we have not disabled
anything. I could be wrong here but I think completely remove SSH
is a major change and it could may some people unhappy.

So for now, we just add the minimum IAM permissions for SSM session.
This is a stripped down version of `AmazonEC2RoleforSSM`.

Signed-off-by: Kai Xia <kaix+github@fastmail.com>